test(core): move diff tests from inline modules to integration tests
The diff tests don't access any crate-private items, so they can run as
integration tests instead. Moving them out drops ~880 lines from the
diff source files and exercises the types through the public API
the same way external consumers do.
